requirements.txt
streamlit
pandas
yfinance
backtrader
datetime
matplotlib
prophet
python-dateutil
requests
torch
numpy
scikit-learn
scipy
torchvision
tqdm
Pillow